+++ packages/components/copybutton/src/CopyButton.test.tsx
@@ -39,6 +39,9 @@ describe('CopyButton', () => {
   });
 
   it('has no a11y issues', async () => {
+    // Workaround for https://github.com/dequelabs/axe-core/issues/3055
+    jest.useRealTimers();
+
     const { container } = render(<CopyButton value="test" />);
     const results = await axe(container);