+++ packages/components/entity-list/src/EntityList.test.tsx
@@ -30,6 +30,9 @@ it('renders the component with an additional class name', () => {
 });
 
 it('has no a11y issues', async () => {
+  // Workaround for https://github.com/dequelabs/axe-core/issues/3055
+  jest.useRealTimers();
+
   const { container } = render(<EntityList />);
   const results = await axe(container);
 
@@ -37,6 +40,9 @@ it('has no a11y issues', async () => {
 });
 
 it('has no a11y issues with children', async () => {
+  // Workaround for https://github.com/dequelabs/axe-core/issues/3055
+  jest.useRealTimers();
+
   const { container } = render(
     <EntityList>
       <EntityList.Item