+++ packages/components/notification/src/NotificationItem.test.tsx
@@ -64,6 +64,9 @@ describe('Notification', () => {
   });
 
   it(`has no a11y issues`, async () => {
+    // Workaround for https://github.com/dequelabs/axe-core/issues/3055
+    jest.useRealTimers();
+
     const mockOnClose = jest.fn();
     const { container } = render(
       <NotificationItem