+++ packages/components/popover/src/Popover.test.tsx
@@ -107,6 +107,7 @@ describe('Popover', function () {
   });
 
   it('call onClose when clicking outside of popover', async () => {
+    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
     const handleClose = jest.fn();
 
     render(
@@ -118,7 +119,7 @@ describe('Popover', function () {
       </Popover>,
     );
 
-    userEvent.click(document.body);
+    await user.click(document.body);
 
     await waitFor(() => {
       expect(handleClose).toHaveBeenCalled();
@@ -126,6 +127,7 @@ describe('Popover', function () {
   });
 
   it('do NOT call onClose when clicking inside of popover', async () => {
+    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
     const handleClose = jest.fn();
 
     render(
@@ -137,7 +139,7 @@ describe('Popover', function () {
       </Popover>,
     );
 
-    userEvent.click(screen.getByRole('dialog'));
+    await user.click(screen.getByRole('dialog'));
 
     await waitFor(() => {
       expect(handleClose).not.toHaveBeenCalled();
@@ -145,6 +147,7 @@ describe('Popover', function () {
   });
 
   it('do NOT call onClose when clicking outside of popover and closeOnBlur is false', async () => {
+    const user = userEvent.setup({ advanceTimers: jest.advanceTimersByTime });
     const handleClose = jest.fn();
 
     render(
@@ -156,7 +159,7 @@ describe('Popover', function () {
       </Popover>,
     );
 
-    userEvent.click(document.body);
+    await user.click(document.body);
 
     await waitFor(() => {
       expect(handleClose).not.toHaveBeenCalled();
@@ -227,6 +230,9 @@ describe('Popover', function () {
   });
 
   it('has no a11y issues', async () => {
+    // Workaround for https://github.com/dequelabs/axe-core/issues/3055
+    jest.useRealTimers();
+
     const { container } = render(
       <Popover isOpen={true}>
         <Popover.Trigger>