+++ packages/core/src/Grid/GridItem/GridItem.test.tsx
@@ -22,6 +22,9 @@ describe('GridItem', () => {
   });
 
   it('has no a11y issues', async () => {
+    // Workaround for https://github.com/dequelabs/axe-core/issues/3055
+    jest.useRealTimers();
+
     const { container } = render(<GridItem>Grid Item</GridItem>);
     const results = await axe(container);