+++ packages/forma-36-codemod/transforms/__testfixtures__/v4/v4-text-field.output.js
@@ -1,4 +1,7 @@
-import { FormControl, TextInput, Flex } from "@contentful/f36-components";
+import { PreviewIcon } from "@contentful/f36-icons";
+
+import { FormControl, TextInput, Flex, Textarea, TextLink } from "@contentful/f36-components";
+
 const {
   id,
   name,
@@ -123,4 +126,30 @@ const validationMessageCondition = 0;
   <FormControl.Label>Conditional validation</FormControl.Label>
   <TextInput name="conditional-validation" onChange={() => {}} />
   {!(validationMessageCondition > 0) && <FormControl.ValidationMessage>Some validation message</FormControl.ValidationMessage>}
+</FormControl>;
+
+<FormControl id="textarea">
+  <FormControl.Label>Text area</FormControl.Label>
+  <Textarea name="textarea" value="some value" rows={5} />
+</FormControl>;
+
+const textLinkProps = {
+  text: 'TextLink',
+  icon: 'Preview',
+  href: '#'
+};
+<FormControl id="with-text-link">
+  <Flex justifyContent="space-between" alignItems="center">
+    <FormControl.Label>With text link</FormControl.Label>
+    <TextLink {...textLinkProps}>{textLinkProps.text}</TextLink>
+  </Flex>
+  <TextInput name="with-text-link" value="some value" />
+</FormControl>;
+
+<FormControl id="with-text-link">
+  <Flex justifyContent="space-between" alignItems="center">
+    <FormControl.Label>With text link</FormControl.Label>
+    <TextLink icon={<PreviewIcon />} href="#">TextLink</TextLink>
+  </Flex>
+  <TextInput name="with-text-link" value="some value" />
 </FormControl>;